backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
mcp>=1.6.0
httpx>=0.26.0
requests>=2.31.0
diff_match_patch
jieba>=0.42.1
pyahocorasick>=2.0.0   # Aho-Corasick automaton for glossary keyword detection
filelock>=3.10.0       # File locking for JSON snapshot concurrency
# Database dependencies
sqlalchemy>=2.0.0
aiosqlite>=0.19.0    # SQLite async driver
asyncpg>=0.29.0      # PostgreSQL async driver
